feat(run-panel): render output with read-only xterm for ANSI + cursor control
Swap the Run panel's plain-text log <div> for a read-only xterm terminal
(the same hook TermPanel uses), so ANSI colors *and* cursor control
(spinners, progress-bar rewrites) render faithfully. Raw PTY bytes from
`run:output` are written straight to the terminal.

- Drop stripAnsi, capLog, the `log` state, the streaming TextDecoder, and
  the auto-scroll effect; xterm now owns decoding, scrollback, and scrolling.
  This also removes the per-chunk React re-render + string realloc.
- Extract resolveTheme/resolveCSSVar/resolveAccentRgba from TermPanel into
  a shared src/util/xtermTheme.ts (used by both panels).
- Add an opt-out `hostOptions.autoFocus` to useXterm so the read-only log
  doesn't steal keyboard focus from the editor.
- Keep last_error as a separate banner (supervisor message, distinct from
  the process stderr already in the terminal).
